requirements.txt
# requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
google-generativeai
pinecone
sentence-transformers
requests
pypdf
python-docx
python-dotenv